Add module plugin to typedoc
