+++ src/common/configuration.ts
