+++ src/common/parallel/parallel-chain.ts
@@ -1,3 +1,8 @@
+/**
+ * @module parallel
+ */
+/** needed, typedoc issue */
+
 import {IParallelStream} from "./parallel-stream";
 import {IEmptyParallelEnvironment, IParallelTaskEnvironment} from "./parallel-environment";
